qbman: Fix 64-bit portability warning(s)
In file included from drivers/qbman/bman_high.c:33:0:
drivers/qbman/bman_low.h: In function ‘RCR_PTR2IDX’:
drivers/qbman/bman_low.h:182:10: warning: cast from pointer to integer of different size

In file included from drivers/qbman/qman_high.c:33:0:
drivers/qbman/qman_low.h: In function ‘EQCR_PTR2IDX’:
drivers/qbman/qman_low.h:282:10: warning: cast from pointer to integer of different size
In file included from drivers/qbman/qman_high.c:33:0:
drivers/qbman/qman_low.h: In function ‘DQRR_PTR2IDX’:
drivers/qbman/qman_low.h:535:10: warning: cast from pointer to integer of different size
In file included from drivers/qbman/qman_high.c:33:0:
drivers/qbman/qman_low.h: In function ‘MR_PTR3IDX’:
drivers/qbman/qman_low.h:852:10: warning: cast from pointer to integer of different size

+++ drivers/qbman/bman_low.h
@@ -4,13 +4,13 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
+ *	 notice, this list of conditions and the following disclaimer.
  *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
+ *	 notice, this list of conditions and the following disclaimer in the
+ *	 documentation and/or other materials provided with the distribution.
  *     * Neither the name of Freescale Semiconductor nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
+ *	 names of its contributors may be used to endorse or promote products
+ *	 derived from this software without specific prior written permission.
  *
  *
  * ALTERNATIVELY, this software may be distributed under the terms of the
@@ -179,7 +179,7 @@ struct bm_portal {
 /* Bit-wise logic to convert a ring pointer to a ring index */
 static inline u8 RCR_PTR2IDX(struct bm_rcr_entry *e)
 {
-	return ((u32)e >> 6) & (BM_RCR_SIZE - 1);
+	return ((uintptr_t)e >> 6) & (BM_RCR_SIZE - 1);
 }
 
 /* Increment the 'cursor' ring pointer, taking 'vbit' into account */
@@ -504,4 +504,3 @@ static inline void __bm_isr_write(struct bm_portal *portal, enum bm_isr_reg n,
 {
 	__bm_out(&portal->addr, REG_ISR + (n << 2), val);
 }
-
